rag agent/requirements.txt
openai>=1.0.0
qdrant-client>=1.9.0
google-generativeai>=0.4.0
cohere>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0